refactor: move code of menu and filters to separate files
